Move current auxv functionality into a new 'auxv' module
